requirements.txt
requests
Flask 
flask-cors 
Pillow 
pytesseract